+++ compat/src/test/java/io/questdb/compat/InfluxDBClientTest.java
@@ -32,7 +32,6 @@ import io.questdb.std.ObjList;
 import io.questdb.std.str.StringSink;
 import io.questdb.test.AbstractBootstrapTest;
 import io.questdb.test.TestServerMain;
-import io.questdb.test.cutlass.http.line.LineHttpUtils;
 import io.questdb.test.tools.TestUtils;
 import org.influxdb.InfluxDB;
 import org.influxdb.dto.BatchPoints;
@@ -47,7 +46,7 @@ import java.util.List;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicReference;
 
-import static io.questdb.test.cutlass.http.line.LineHttpUtils.assertRequestErrorContains;
+import static io.questdb.compat.InfluxDBUtils.assertRequestErrorContains;
 
 public class InfluxDBClientTest extends AbstractBootstrapTest {
 
@@ -68,7 +67,7 @@ public class InfluxDBClientTest extends AbstractBootstrapTest {
                 serverMain.compile("create table ex_tbl(b byte, s short, f float, d double, str string, sym symbol, tss timestamp, " +
                         "i int, l long, ip ipv4, g geohash(4c), ts timestamp) timestamp(ts) partition by DAY WAL");
 
-                try (final InfluxDB influxDB = LineHttpUtils.getConnection(serverMain)) {
+                try (final InfluxDB influxDB = InfluxDBUtils.getConnection(serverMain)) {
                     List<String> points = new ArrayList<>();
 
                     assertRequestErrorContains(influxDB, points, "ex_tbl b\\\"c=1024 1233456\n", "{" +
@@ -119,7 +118,7 @@ public class InfluxDBClientTest extends AbstractBootstrapTest {
 
                 // New column added
                 lines.add("wal_low_max_uncomitted i=123i\n");
-                try (final InfluxDB influxDB = LineHttpUtils.getConnection(serverMain)) {
+                try (final InfluxDB influxDB = InfluxDBUtils.getConnection(serverMain)) {
                     // Bad line which should roll back the transaction
                     assertRequestErrorContains(influxDB, lines, "ailed to parse line protocol:errors encountered on line(s):" +
                             "\\nerror in line 10002: Could not parse entire line. Symbol value is missing: bla");
@@ -141,7 +140,7 @@ public class InfluxDBClientTest extends AbstractBootstrapTest {
                 serverMain.compile("create table wal_not_here(b byte, s short, f float, d double, str string, sym symbol, tss timestamp, " +
                         "i int, l long, ip ipv4, g geohash(4c), ts timestamp)");
 
-                try (final InfluxDB influxDB = LineHttpUtils.getConnection(serverMain)) {
+                try (final InfluxDB influxDB = InfluxDBUtils.getConnection(serverMain)) {
 
                     List<String> points = new ArrayList<>();
                     assertRequestErrorContains(influxDB, points, "badPo\"int,a3=2 1233456\n", "{" +
@@ -195,8 +194,7 @@ public class InfluxDBClientTest extends AbstractBootstrapTest {
                 serverMain.compile("create table wal_not_here(b byte, s short, f float, d double, str string, sym symbol, tss timestamp, " +
                         "i int, l long, ip ipv4, g geohash(4c), ts timestamp)");
 
-                try (final InfluxDB influxDB = LineHttpUtils.getConnection(serverMain)) {
-
+                try (final InfluxDB influxDB = InfluxDBUtils.getConnection(serverMain)) {
                     List<String> points = new ArrayList<>();
 
                     StringSink sink = new StringSink();
@@ -360,8 +358,7 @@ public class InfluxDBClientTest extends AbstractBootstrapTest {
                 serverMain.compile("create table wal_not_here(b byte, s short, f float, d double, str string, sym symbol, tss timestamp, " +
                         "i int, l long, ip ipv4, g geohash(4c), ts timestamp)");
 
-                try (final InfluxDB influxDB = LineHttpUtils.getConnection(serverMain)) {
-
+                try (final InfluxDB influxDB = InfluxDBUtils.getConnection(serverMain)) {
                     List<String> points = new ArrayList<>();
 
                     // Fail on first line
@@ -411,7 +408,7 @@ public class InfluxDBClientTest extends AbstractBootstrapTest {
                     PropertyKey.HTTP_RECEIVE_BUFFER_SIZE.getEnvVarName(), "2048"
             )) {
                 serverMain.start();
-                try (final InfluxDB influxDB = LineHttpUtils.getConnection(serverMain)) {
+                try (final InfluxDB influxDB = InfluxDBUtils.getConnection(serverMain)) {
 
                     List<String> points = new ArrayList<>();
                     points.add("good_point,sym=a str=\"abdc\",num=1 1233456\n");
@@ -487,8 +484,7 @@ public class InfluxDBClientTest extends AbstractBootstrapTest {
                 serverMain.compile("create table wal_not_here(b byte, s short, f float, d double, str string, sym symbol, tss timestamp, " +
                         "i int, l long, ip ipv4, g geohash(4c), ts timestamp)");
 
-                try (final InfluxDB influxDB = LineHttpUtils.getConnection(serverMain)) {
-
+                try (final InfluxDB influxDB = InfluxDBUtils.getConnection(serverMain)) {
                     List<String> points = new ArrayList<>();
                     points.add("m1,tag1=value1 f1=1i,y=12i");
                     points.add("m1,tag1=value1 f1=1i,x=12i");
@@ -508,7 +504,7 @@ public class InfluxDBClientTest extends AbstractBootstrapTest {
             )) {
                 serverMain.start();
 
-                try (final InfluxDB influxDB = LineHttpUtils.getConnection(serverMain)) {
+                try (final InfluxDB influxDB = InfluxDBUtils.getConnection(serverMain)) {
                     influxDB.setLogLevel(InfluxDB.LogLevel.FULL);
                     Pong pong = influxDB.ping();
                     Assert.assertTrue(pong.isGood());
@@ -535,7 +531,7 @@ public class InfluxDBClientTest extends AbstractBootstrapTest {
                     lines.add(goodLine);
                 }
 
-                try (final InfluxDB influxDB = LineHttpUtils.getConnection(serverMain)) {
+                try (final InfluxDB influxDB = InfluxDBUtils.getConnection(serverMain)) {
                     // Bad line which should roll back the transaction
                     int totalCount = count + 1;
                     assertRequestErrorContains(influxDB, lines, "wal_low_max_uncomitted,bla i=aaa\n",
@@ -573,7 +569,7 @@ public class InfluxDBClientTest extends AbstractBootstrapTest {
                     lines.add(addColumnLine);
                 }
 
-                try (final InfluxDB influxDB = LineHttpUtils.getConnection(serverMain)) {
+                try (final InfluxDB influxDB = InfluxDBUtils.getConnection(serverMain)) {
                     // Column is added
                     influxDB.write(lines);
                 }
@@ -595,8 +591,7 @@ public class InfluxDBClientTest extends AbstractBootstrapTest {
                 serverMain.compile("create table ex_tbl(b byte, s short, f float, d double, str string, sym symbol, tss timestamp, " +
                         "i int, l long, ip ipv4, g geohash(4c), ts timestamp) timestamp(ts) partition by DAY WAL");
 
-                try (final InfluxDB influxDB = LineHttpUtils.getConnection(serverMain)) {
-
+                try (final InfluxDB influxDB = InfluxDBUtils.getConnection(serverMain)) {
                     List<String> points = new ArrayList<>();
                     assertRequestErrorContains(influxDB, points, "ex_tbl,a3=2 1222233456\n", "{" +
                             "\"code\":\"invalid\"," +
@@ -625,8 +620,7 @@ public class InfluxDBClientTest extends AbstractBootstrapTest {
                 serverMain.compile("create table ex_tbl(b byte, s short, f float, d double, str string, sym symbol, tss timestamp, " +
                         "i int, l long, ip ipv4, g geohash(4c), ts timestamp) timestamp(ts) partition by DAY WAL");
 
-                try (final InfluxDB influxDB = LineHttpUtils.getConnection(serverMain)) {
-
+                try (final InfluxDB influxDB = InfluxDBUtils.getConnection(serverMain)) {
                     List<String> points = new ArrayList<>();
                     assertRequestErrorContains(influxDB, points, "ex_tbl,a3=2 1222233456\n", "{" +
                             "\"code\":\"invalid\"," +
@@ -651,7 +645,7 @@ public class InfluxDBClientTest extends AbstractBootstrapTest {
             )) {
                 serverMain.start();
 
-                try (final InfluxDB influxDB = LineHttpUtils.getConnection(serverMain)) {
+                try (final InfluxDB influxDB = InfluxDBUtils.getConnection(serverMain)) {
                     influxDB.setLogLevel(InfluxDB.LogLevel.BASIC);
 
                     long microTime = IntervalUtils.parseFloorPartialTimestamp("2022-02-24T04:00:00.000001Z");
@@ -704,8 +698,7 @@ public class InfluxDBClientTest extends AbstractBootstrapTest {
         long timestamp = IntervalUtils.parseFloorPartialTimestamp("2023-11-27T18:53:24.834Z");
         int i = 0;
 
-        try (final InfluxDB influxDB = LineHttpUtils.getConnection(serverMain)) {
-
+        try (final InfluxDB influxDB = InfluxDBUtils.getConnection(serverMain)) {
             BatchPoints batchPoints = BatchPoints
                     .database("test_db")
                     .tag("async", "true")
