+++ core/src/test/java/module-info.java
@@ -30,9 +30,6 @@ open module io.questdb.test {
     requires static java.sql;
     requires static org.postgresql.jdbc;
     requires static java.management;
-    requires influxdb.java;
-    requires okhttp3;
-    requires okio;
 
     uses io.questdb.griffin.FunctionFactory;
